backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "openai>=2.32.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.12",
    "pymupdf>=1.27.2.3",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "tiktoken>=0.12.0",
    "uvicorn>=0.46.0",
]
